+++ sdk/datafactory/Azure.ResourceManager.DataFactory/tests/Scenario/DataFactoryDatasetResourceTests.cs
@@ -284,7 +284,6 @@ namespace Azure.ResourceManager.DataFactory.Tests.Scenario
 
         [Test]
         [RecordedTest]
-        [Ignore("test issue")]
         public async Task Dataset_OracleTable_Create()
         {
             await DatasetCreate("oracle", CreateOracleLinkedService, (string linkedServiceName) =>