+++ sdk/operationalinsights/Azure.ResourceManager.OperationalInsights/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.2.0-beta.2 (Unreleased)
+## 1.3.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -12,6 +12,16 @@
 - Add `201` to the response code of the `QueryPacks_UpdateTags`.
 
 ### Other Changes
+
+## 1.2.0 (2023-11-29)
+
+### Features Added
+
+- Enable mocking for extension methods, refer this [document](https://aka.ms/azsdk/net/mocking) for more details.
+
+### Other Changes
+
+- Upgraded dependent `Azure.ResourceManager` to 1.9.0.
 
 ## 1.2.0-beta.1 (2023-05-31)